Allow DataStore#upload to support all put methods
Per the documentation in GeoServer, creating a datastore via the REST
API is possibly not only by uploading a file, but also by referencing a
remote URL, or an existing (absolute) reference to a file on the
Geoserver.

see: https://docs.geoserver.org/stable/en/api/#1.0.0/datastores.yaml

This commit aims to support all 3 documented options by introducing the
`method` parameter to `DataStore#upload` and using a default value of
`file` to match the previous default implementation.